requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
# -e git+https://github.com/openai/baselines.git@ea25b9e8b234e6ee1bca43083f8f3cf974143998#egg=baselines
box2d-py==2.3.8
cffi==1.14.0
click==7.1.2
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.21
EasyProcess==0.3
fasteners==0.15
future==0.18.2
gast==0.3.3
glfw==1.12.0
google-pasta==0.2.0
grpcio==1.30.0
gym==0.15.7
h5py==2.10.0
imageio==2.9.0
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.2.2
matplotlib==3.2.2
monotonic==1.5
# mujoco-py==1.50.1.68
numpy==1.19.0
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
protobuf==3.12.2
pycparser==2.20
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyVirtualDisplay==1.3.2
scipy==1.5.1
six==1.15.0
stable-baselines==2.10.0
tensorflow==1.14.0
termcolor==1.1.0
torch==1.5.1
tqdm==4.47.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
